完成附近周边店铺搜索功能

+++ app/Product.php
@@ -49,7 +49,7 @@ class Product extends Model
                 'type'       => 'nested',
                 'properties' => [
                     'id'   => [
-                        'type' => 'long',
+                        'type' => 'short',
                     ],
                     'name' => [
                         'type' => 'text',
